# Conflicts:
#	README.md
#	app/electron/main.js
#	app/package.json
#	app/pnpm-lock.yaml
#	app/src/block/popover.ts
#	app/src/boot/globalEvent/keydown.ts
#	app/src/boot/onGetConfig.ts
#	app/src/card/viewCards.ts
#	app/src/config/editor.ts
#	app/src/config/publish.ts
#	app/src/config/repos.ts
#	app/src/config/search.ts
#	app/src/editor/deleteFile.ts
#	app/src/history/diff.ts
#	app/src/history/history.ts
#	app/src/index.ts
#	app/src/layout/dock/Backlink.ts
#	app/src/layout/dock/Bookmark.ts
#	app/src/layout/dock/Files.ts
#	app/src/layout/dock/Graph.ts
#	app/src/layout/dock/Outline.ts
#	app/src/layout/dock/Tag.ts
#	app/src/layout/getAll.ts
#	app/src/menus/commonMenuItem.ts
#	app/src/menus/navigation.ts
#	app/src/menus/protyle.ts
#	app/src/menus/util.ts
#	app/src/mobile/dock/MobileFiles.ts
#	app/src/plugin/API.ts
#	app/src/plugin/uninstall.ts
#	app/src/protyle/gutter/index.ts
#	app/src/protyle/header/Background.ts
#	app/src/protyle/header/openTitleMenu.ts
#	app/src/protyle/index.ts
#	app/src/protyle/preview/image.ts
#	app/src/protyle/render/abcRender.ts
#	app/src/protyle/render/av/asset.ts
#	app/src/protyle/render/av/blockAttr.ts
#	app/src/protyle/render/av/cell.ts
#	app/src/protyle/render/av/openMenuPanel.ts
#	app/src/protyle/render/blockRender.ts
#	app/src/protyle/render/chartRender.ts
#	app/src/protyle/render/graphvizRender.ts
#	app/src/protyle/render/htmlRender.ts
#	app/src/protyle/render/mathRender.ts
#	app/src/protyle/render/mermaidRender.ts
#	app/src/protyle/toolbar/Link.ts
#	app/src/protyle/toolbar/index.ts
#	app/src/protyle/upload/index.ts
#	app/src/protyle/util/editorCommonEvent.ts
#	app/src/protyle/util/paste.ts
#	app/src/protyle/util/processCode.ts
#	app/src/protyle/wysiwyg/index.ts
#	app/src/protyle/wysiwyg/transaction.ts
#	app/src/util/escape.ts
#	app/src/window/openNewWindow.ts
#	kernel/go.mod
#	kernel/sql/block_ref_query.go
#	kernel/util/working.go

+++ kernel/api/import.go
@@ -17,6 +17,7 @@
 package api
 
 import (
+	"fmt"
 	"io"
 	"net/http"
 	"os"
@@ -53,14 +54,6 @@ func importSY(c *gin.Context) {
 		return
 	}
 	file := files[0]
-	reader, err := file.Open()
-	if err != nil {
-		logging.LogErrorf("read import .sy.zip failed: %s", err)
-		ret.Code = -1
-		ret.Msg = err.Error()
-		return
-	}
-
 	importDir := filepath.Join(util.TempDir, "import")
 	if err = os.MkdirAll(importDir, 0755); err != nil {
 		logging.LogErrorf("make import dir [%s] failed: %s", importDir, err)
@@ -68,9 +61,37 @@ func importSY(c *gin.Context) {
 		ret.Msg = err.Error()
 		return
 	}
-	writePath := filepath.Join(util.TempDir, "import", file.Filename)
+
+	writePath := filepath.Join(importDir, file.Filename)
+	if !gulu.File.IsSubPath(importDir, writePath) {
+		logging.LogErrorf("import path [%s] is not sub path of import dir [%s]", writePath, importDir)
+		ret.Code = -1
+		ret.Msg = "import path is not sub path of import dir"
+		return
+	}
+
 	defer os.RemoveAll(writePath)
-	writer, err := os.OpenFile(writePath, os.O_RDWR|os.O_CREATE, 0644)
+
+	var reader io.ReadCloser
+	var writer *os.File
+	defer func() {
+		if writer != nil {
+			_ = writer.Close()
+		}
+		if reader != nil {
+			_ = reader.Close()
+		}
+	}()
+
+	reader, err = file.Open()
+	if err != nil {
+		logging.LogErrorf("read import .sy.zip failed: %s", err)
+		ret.Code = -1
+		ret.Msg = err.Error()
+		return
+	}
+
+	writer, err = os.OpenFile(writePath, os.O_RDWR|os.O_CREATE, 0644)
 	if err != nil {
 		logging.LogErrorf("open import .sy.zip [%s] failed: %s", writePath, err)
 		ret.Code = -1
@@ -83,8 +104,20 @@ func importSY(c *gin.Context) {
 		ret.Msg = err.Error()
 		return
 	}
-	writer.Close()
-	reader.Close()
+	if err = writer.Close(); err != nil {
+		logging.LogErrorf("close import .sy.zip [%s] failed: %s", writePath, err)
+		ret.Code = -1
+		ret.Msg = err.Error()
+		return
+	}
+	writer = nil
+	if err = reader.Close(); err != nil {
+		logging.LogErrorf("close import upload reader failed: %s", err)
+		ret.Code = -1
+		ret.Msg = err.Error()
+		return
+	}
+	reader = nil
 
 	notebook := form.Value["notebook"][0]
 	toPath := form.Value["toPath"][0]
@@ -119,16 +152,28 @@ func importData(c *gin.Context) {
 		return
 	}
 
-	tmpImport := filepath.Join(util.TempDir, "import")
-	err = os.MkdirAll(tmpImport, 0755)
+	importDir := filepath.Join(util.TempDir, "import")
+	err = os.MkdirAll(importDir, 0755)
 	if err != nil {
 		ret.Code = -1
 		ret.Msg = "create temp import dir failed"
 		return
 	}
-	dataZipPath := filepath.Join(tmpImport, util.CurrentTimeSecondsStr()+".zip")
+	dataZipPath := filepath.Join(importDir, util.CurrentTimeSecondsStr()+".zip")
 	defer os.RemoveAll(dataZipPath)
-	dataZipFile, err := os.Create(dataZipPath)
+
+	var dataZipFile *os.File
+	var fileReader io.ReadCloser
+	defer func() {
+		if dataZipFile != nil {
+			_ = dataZipFile.Close()
+		}
+		if fileReader != nil {
+			_ = fileReader.Close()
+		}
+	}()
+
+	dataZipFile, err = os.Create(dataZipPath)
 	if err != nil {
 		logging.LogErrorf("create temp file failed: %s", err)
 		ret.Code = -1
@@ -137,7 +182,7 @@ func importData(c *gin.Context) {
 	}
 	file := form.File["file"][0]
 	logging.LogInfof("import data [name=%s, size=%d]", file.Filename, file.Size)
-	fileReader, err := file.Open()
+	fileReader, err = file.Open()
 	if err != nil {
 		logging.LogErrorf("open upload file failed: %s", err)
 		ret.Code = -1
@@ -157,7 +202,14 @@ func importData(c *gin.Context) {
 		ret.Msg = "close file failed"
 		return
 	}
-	fileReader.Close()
+	dataZipFile = nil
+	if err = fileReader.Close(); err != nil {
+		logging.LogErrorf("close upload reader failed: %s", err)
+		ret.Code = -1
+		ret.Msg = "close file failed"
+		return
+	}
+	fileReader = nil
 
 	err = model.ImportData(dataZipPath)
 	if err != nil {
@@ -179,6 +231,23 @@ func importStdMd(c *gin.Context) {
 	notebook := arg["notebook"].(string)
 	localPath := arg["localPath"].(string)
 	toPath := arg["toPath"].(string)
+
+	if gulu.File.IsSubPath(util.WorkingDir, localPath) {
+		msg := fmt.Sprintf("import from local path [%s] failed: local path is sub path of working dir", localPath)
+		logging.LogErrorf(msg)
+		ret.Code = -1
+		ret.Msg = msg
+		return
+	}
+
+	if util.IsSensitivePath(localPath) {
+		msg := fmt.Sprintf("import from local path [%s] failed: local path is sensitive path", localPath)
+		logging.LogErrorf(msg)
+		ret.Code = -1
+		ret.Msg = msg
+		return
+	}
+
 	err := model.ImportFromLocalPath(notebook, localPath, toPath)
 	if err != nil {
 		ret.Code = -1
@@ -210,14 +279,6 @@ func importZipMd(c *gin.Context) {
 		return
 	}
 	file := files[0]
-	reader, err := file.Open()
-	if err != nil {
-		logging.LogErrorf("read import .zip failed: %s", err)
-		ret.Code = -1
-		ret.Msg = err.Error()
-		return
-	}
-
 	importDir := filepath.Join(util.TempDir, "import")
 	if err = os.MkdirAll(importDir, 0755); err != nil {
 		logging.LogErrorf("make import dir [%s] failed: %s", importDir, err)
@@ -225,9 +286,37 @@ func importZipMd(c *gin.Context) {
 		ret.Msg = err.Error()
 		return
 	}
-	writePath := filepath.Join(util.TempDir, "import", file.Filename)
+
+	writePath := filepath.Join(importDir, file.Filename)
+	if !gulu.File.IsSubPath(importDir, writePath) {
+		logging.LogErrorf("import path [%s] is not sub path of import dir [%s]", writePath, importDir)
+		ret.Code = -1
+		ret.Msg = "import path is not sub path of import dir"
+		return
+	}
+
 	defer os.RemoveAll(writePath)
-	writer, err := os.OpenFile(writePath, os.O_RDWR|os.O_CREATE, 0644)
+
+	var reader io.ReadCloser
+	var writer *os.File
+	defer func() {
+		if writer != nil {
+			_ = writer.Close()
+		}
+		if reader != nil {
+			_ = reader.Close()
+		}
+	}()
+
+	reader, err = file.Open()
+	if err != nil {
+		logging.LogErrorf("read import .zip failed: %s", err)
+		ret.Code = -1
+		ret.Msg = err.Error()
+		return
+	}
+
+	writer, err = os.OpenFile(writePath, os.O_RDWR|os.O_CREATE, 0644)
 	if err != nil {
 		logging.LogErrorf("open import .zip [%s] failed: %s", writePath, err)
 		ret.Code = -1
@@ -240,8 +329,20 @@ func importZipMd(c *gin.Context) {
 		ret.Msg = err.Error()
 		return
 	}
-	writer.Close()
-	reader.Close()
+	if err = writer.Close(); err != nil {
+		logging.LogErrorf("close import .zip [%s] failed: %s", writePath, err)
+		ret.Code = -1
+		ret.Msg = err.Error()
+		return
+	}
+	writer = nil
+	if err = reader.Close(); err != nil {
+		logging.LogErrorf("close import upload reader failed: %s", err)
+		ret.Code = -1
+		ret.Msg = err.Error()
+		return
+	}
+	reader = nil
 
 	notebook := form.Value["notebook"][0]
 	toPath := form.Value["toPath"][0]
